A little trouble with this code
I'm having a little trouble with this code here. From line 22 to around line 32, It's been getting stuck in a loop of constantly asking to re-enter a password with numbers/letters included depending on which the password doesn't have. It's fine if the password has letters and numbers, but if it doesn't have one, the loop occurs. I was wondering if I could possibly get some help with this? Thanks :)